python/requirements.txt
stereovision
opencv-contrib-python
matplotlib
pyaml